Adds the function are_approximately_equal
This adds the function approximately equal to compare two
instances (tensors) for shape, allocator attributes, and
(up to a specified precision) elements. This may be
necessary in general since compiler optimizations may mean
that the initial state, order of
operations, compilation proedure and hardware execution
cannot be guaranteed (see issue #51).

* Adds two files /include/tensorwrapper/tensor/approximately_equal.hpp
  and /src/tensorwrapper/tensor/approximately_equal.cpp to add the
  are_approximately_equal function at the TW level.

* Modifies five files /include/tensorwrapper/tensor/buffer/buffer.cpp,
  /src/tensorwrapper/tensor/buffer/buffer.hpp,
  /src/tensorwrapper/tensor/buffer/detail_/buffer_pimpl.hpp,
  /src/tensorwrapper/tensor/buffer/detail_/ta_buffer_pimpl.hpp,
  and /src/tensorwrapper/tensor/buffer/detail_/ta_buffer_pimpl.cpp
  to implement the equivalent at the Buffer level