Adds function are_approximately_equal
This function compares two instances of TW for equality of elements,
to within a specified tolerance, and for equality

The file /include/tensorwrapper/tensor/approximately_equal.hpp
contains the declaration and definition of the function
"are_approximately_equal" which compares two TW instances.

The file /include/tensorwrapper/tensor/tensor.hpp has been
modified to include a reference to the new function

The file /tests/tensor/are_approximately_equal.cpp is
a minimal test for the new function.